# perf: use `@jridgewell/sourcemap-codec`