[`@jridgewell/sourcemap-codec` has better performance and memory footprint than `sourcemap-codec`](https://github.com/jridgewell/sourcemap-codec#benchmarks).

This PR replaces `sourcemap-codec` with `@jridgewell/sourcemap-codec`.

**before**
```
node v16.16.0

generateMap (no edit) x 1,664 ops/sec ±0.62% (92 runs sampled)
generateMap (edit) x 1,651 ops/sec ±0.74% (92 runs sampled)
```
**after**
```
node v16.16.0

generateMap (no edit) x 1,981 ops/sec ±0.52% (93 runs sampled)
generateMap (edit) x 1,901 ops/sec ±0.56% (93 runs sampled)
```