@sapphi-red 2022-07-07
@aleclarson FYI this pull request is waiting to see what happens with https://github.com/Rich-Harris/magic-string/pull/216#issuecomment-1146691084.