@jdalton 2022-08-12
@jridgewell it looks like `@jridgewell/sourcemap-codec` is a fork of Rich's. Do you have a summary of changed done to it?

To others: Should the original `sourcemap-codec` be deprecated?