@jridgewell 2022-08-12
You're only using the `encode` methods, to the relevant change is the use of a TextDecoder (or Buffer if your node version is old) to convert from the base64 bytes into a the string. 

Sourcemaps are essentially a `Line[]`, where each Line` is a `Segment[]`, and each Segment is a tuple of ints. Rich's implementations allocates an array per line, encodes each segment into a base64 string (using small concatenations, which is too small for v8's rope concatenations to kick in), and pushes the encoding into the line (array expansion costs). It then joins the array together, and pushes that onto the overall mappings string (this uses rope optimizations one the mappings is large enough).

My implementation uses a 16kb Uint8Array buffer. The individual base64 char is placed into the buffer, and when it fills I use a TextDecoder to covert them into a string in one shot. I then start filling the buffer again. This avoids small string concatenations, array expansions, and guarantees the rope optimizations kick in immediately (because 16kb is large enough).