@Rich-Harris 2022-12-03
> To others: Should the original `sourcemap-codec` be deprecated?

Yes, I think it should (sorry, dropped the ball on this whole convo, doing some inbox spring-cleaning). The fork has more downloads than the original, and I would wager the vast majority of `sourcemap-codec`'s downloads are via `magic-string`, so deprecating it in favour of `@jridgewell/sourcemap-codec` would be the least disruptive move. Thanks @jridgewell for taking the initiative, I appreciate it 🍻