@Rich-Harris 2022-12-03
Released 0.27 with this change